Fixed bug caused by an Integer passed instead of a Long.
git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@7017 deaff617-5be1-0310-a1b7-ec6878f92ba9

+++ applications/product/src/org/ofbiz/shipment/shipment/ShipmentServices.java
@@ -584,7 +584,7 @@ public class ShipmentServices {
             stageShip.set("postalCodeExt", address.get("postalCodeExt"));
             stageShip.set("countryGeoId", address.get("countryGeoId"));
             stageShip.set("stateProvinceGeoId", address.get("stateProvinceGeoId"));
-            stageShip.set("numberOfPackages", new Integer(packages.size()));
+            stageShip.set("numberOfPackages", new Long(packages.size()));
             stageShip.set("handlingInstructions", shipment.get("handlingInstructions"));
             toStore.add(stageShip);
